Remove extra files
