+++ main.py
@@ -3,14 +3,14 @@ import sys
 from qtpy import QtWidgets, QtGui, QtCore
 from qtpy.QtGui import QIcon
 
-from pyqtribbon import RibbonBar, RibbonCategoryStyle, Large, RibbonStyle, RibbonMainWindow
+from pyqtribbon import RibbonBar, RibbonCategoryStyle, Large, RibbonStyle
 
 if __name__ == "__main__":
     app = QtWidgets.QApplication(sys.argv)
     app.setFont(QtGui.QFont("Times New Roman", 10))
     app.setStyle("Windows")
 
-    window = RibbonMainWindow()
+    window = QtWidgets.QMainWindow()
     ribbon = RibbonBar(maxRows=6)
     ribbon.setRibbonStyle(RibbonStyle.Default)
     window.setMenuBar(ribbon)
