+++ pyqtribbon/__init__.py
@@ -4,6 +4,5 @@ from .menu import RibbonMenu, RibbonPermanentMenu
 from .panel import RibbonSpaceFindMode, ColumnWise, RowWise
 from .ribbonbar import RibbonBar, RibbonStyle, Default, Debug
 from .toolbutton import RibbonButtonStyle, Small, Medium, Large
-from .mainwindow import RibbonMainWindow
 
 _version__ = "0.4.0"
