--- pyqtribbon/framelesswindow/rc/resource.qrc
@@ -1,6 +1,0 @@
-<RCC>
-    <qresource prefix="/framelesswindow">
-        <file alias="close_black.svg">title_bar/close_black.svg</file>
-        <file alias="close_white.svg">title_bar/close_white.svg</file>
-    </qresource>
-</RCC>
