--- pyqtribbon/framelesswindow/unix/__init__.py
@@ -1,77 +1,0 @@
-# coding:utf-8
-from PyQt5.QtCore import QCoreApplication, QEvent, Qt
-from PyQt5.QtGui import QMouseEvent
-from PyQt5.QtWidgets import QMainWindow
-
-from ..titlebar import TitleBar
-from ..utils.linux_utils import LinuxMoveResize
-from .window_effect import UnixWindowEffect
-
-
-class UnixFramelessWindow(QMainWindow):
-    """ Frameless window for Unix system """
-
-    BORDER_WIDTH = 5
-
-    def __init__(self, parent=None):
-        super().__init__(parent=parent)
-        self.windowEffect = UnixWindowEffect()
-        self.titleBar = TitleBar(self)
-
-        self.setWindowFlags(self.windowFlags() | Qt.FramelessWindowHint)
-        QCoreApplication.instance().installEventFilter(self)
-
-        self.titleBar.raise_()
-        self.resize(500, 500)
-        self.setContentsMargins(0, self.titleBar.height(), 0, 0)
-
-    def resizeEvent(self, e):
-        super().resizeEvent(e)
-        self.titleBar.resize(self.width(), self.titleBar.height())
-
-    def setTitleBar(self, titleBar):
-        """ set custom title bar
-
-        Parameters
-        ----------
-        titleBar: TitleBar
-            title bar
-        """
-        self.titleBar.deleteLater()
-        self.titleBar = titleBar
-        self.titleBar.setParent(self)
-        self.titleBar.raise_()
-
-    def eventFilter(self, obj, event):
-        et = event.type()
-        if et not in [QEvent.MouseButtonPress, QEvent.MouseMove]:
-            return False
-
-        edges = Qt.Edges()
-        pos = QMouseEvent(event).globalPos() - self.pos()
-        if pos.x() < self.BORDER_WIDTH:
-            edges |= Qt.LeftEdge
-        if pos.x() >= self.width()-self.BORDER_WIDTH:
-            edges |= Qt.RightEdge
-        if pos.y() < self.BORDER_WIDTH:
-            edges |= Qt.TopEdge
-        if pos.y() >= self.height()-self.BORDER_WIDTH:
-            edges |= Qt.BottomEdge
-
-        # change cursor
-        if et == QEvent.MouseMove and self.windowState() == Qt.WindowNoState:
-            if edges in (Qt.LeftEdge | Qt.TopEdge, Qt.RightEdge | Qt.BottomEdge):
-                self.setCursor(Qt.SizeFDiagCursor)
-            elif edges in (Qt.RightEdge | Qt.TopEdge, Qt.LeftEdge | Qt.BottomEdge):
-                self.setCursor(Qt.SizeBDiagCursor)
-            elif edges in (Qt.TopEdge, Qt.BottomEdge):
-                self.setCursor(Qt.SizeVerCursor)
-            elif edges in (Qt.LeftEdge, Qt.RightEdge):
-                self.setCursor(Qt.SizeHorCursor)
-            else:
-                self.setCursor(Qt.ArrowCursor)
-
-        elif obj in (self, self.titleBar) and et == QEvent.MouseButtonPress and edges:
-            LinuxMoveResize.starSystemResize(self, event.globalPos(), edges)
-
-        return super().eventFilter(obj, event)
