--- pyqtribbon/framelesswindow/windows/__init__.py
@@ -1,173 +1,0 @@
-# coding:utf-8
-import sys
-from ctypes import cast
-from ctypes.wintypes import LPRECT, MSG
-from platform import platform
-
-import win32con
-import win32gui
-from PyQt5.QtCore import Qt
-from PyQt5.QtGui import QCloseEvent, QCursor
-from PyQt5.QtWidgets import QApplication, QMainWindow
-from PyQt5.QtWinExtras import QtWin
-
-from ..titlebar import TitleBar
-from ..utils import win32_utils as win_utils
-from ..utils.win32_utils import Taskbar
-from .c_structures import LPNCCALCSIZE_PARAMS
-from .window_effect import WindowsWindowEffect
-
-
-class WindowsFramelessWindow(QMainWindow):
-    """  Frameless window for Windows system """
-
-    BORDER_WIDTH = 5
-
-    def __init__(self, parent=None):
-        super().__init__(parent=parent)
-        self.windowEffect = WindowsWindowEffect()
-        self.titleBar = TitleBar(self)
-
-        # remove window border
-        self.setWindowFlags(self.windowFlags() | Qt.FramelessWindowHint)
-
-        # add DWM shadow and window animation
-        self.windowEffect.addWindowAnimation(self.winId())
-        if not isinstance(self, AcrylicWindow):
-            self.windowEffect.addShadowEffect(self.winId())
-
-        # solve issue #5
-        self.windowHandle().screenChanged.connect(self.__onScreenChanged)
-
-        self.resize(500, 500)
-        self.titleBar.raise_()
-        self.setContentsMargins(0, self.titleBar.height(), 0, 0)
-
-    def setTitleBar(self, titleBar):
-        """ set custom title bar
-
-        Parameters
-        ----------
-        titleBar: TitleBar
-            title bar
-        """
-        self.titleBar.deleteLater()
-        self.titleBar = titleBar
-        self.titleBar.setParent(self)
-        self.titleBar.raise_()
-
-    def resizeEvent(self, e):
-        super().resizeEvent(e)
-        self.titleBar.resize(self.width(), self.titleBar.height())
-
-    def nativeEvent(self, eventType, message):
-        """ Handle the Windows message """
-        msg = MSG.from_address(message.__int__())
-        if not msg.hWnd:
-            return super().nativeEvent(eventType, message)
-
-        if msg.message == win32con.WM_NCHITTEST:
-            pos = QCursor.pos()
-            xPos = pos.x() - self.x()
-            yPos = pos.y() - self.y()
-            w, h = self.width(), self.height()
-            lx = xPos < self.BORDER_WIDTH
-            rx = xPos > w - self.BORDER_WIDTH
-            ty = yPos < self.BORDER_WIDTH
-            by = yPos > h - self.BORDER_WIDTH
-            if lx and ty:
-                return True, win32con.HTTOPLEFT
-            elif rx and by:
-                return True, win32con.HTBOTTOMRIGHT
-            elif rx and ty:
-                return True, win32con.HTTOPRIGHT
-            elif lx and by:
-                return True, win32con.HTBOTTOMLEFT
-            elif ty:
-                return True, win32con.HTTOP
-            elif by:
-                return True, win32con.HTBOTTOM
-            elif lx:
-                return True, win32con.HTLEFT
-            elif rx:
-                return True, win32con.HTRIGHT
-        elif msg.message == win32con.WM_NCCALCSIZE:
-            if msg.wParam:
-                rect = cast(msg.lParam, LPNCCALCSIZE_PARAMS).contents.rgrc[0]
-            else:
-                rect = cast(msg.lParam, LPRECT).contents
-
-            isMax = win_utils.isMaximized(msg.hWnd)
-            isFull = win_utils.isFullScreen(msg.hWnd)
-
-            # adjust the size of client rect
-            if isMax and not isFull:
-                thickness = win_utils.getResizeBorderThickness(msg.hWnd)
-                rect.top += thickness
-                rect.left += thickness
-                rect.right -= thickness
-                rect.bottom -= thickness
-
-            # handle the situation that an auto-hide taskbar is enabled
-            if (isMax or isFull) and Taskbar.isAutoHide():
-                position = Taskbar.getPosition(msg.hWnd)
-                if position == Taskbar.LEFT:
-                    rect.top += Taskbar.AUTO_HIDE_THICKNESS
-                elif position == Taskbar.BOTTOM:
-                    rect.bottom -= Taskbar.AUTO_HIDE_THICKNESS
-                elif position == Taskbar.RIGHT:
-                    rect.right -= Taskbar.AUTO_HIDE_THICKNESS
-
-            result = win32con.WVR_REDRAW if msg.wParam else 0
-            return True, result
-
-        return super().nativeEvent(eventType, message)
-
-    def __onScreenChanged(self):
-        hWnd = int(self.windowHandle().winId())
-        win32gui.SetWindowPos(hWnd, None, 0, 0, 0, 0, win32con.SWP_NOMOVE |
-                              win32con.SWP_NOSIZE | win32con.SWP_FRAMECHANGED)
-
-
-class AcrylicWindow(WindowsFramelessWindow):
-    """ A frameless window with acrylic effect """
-
-    def __init__(self, parent=None):
-        super().__init__(parent=parent)
-        self.__closedByKey = False
-
-        QtWin.enableBlurBehindWindow(self)
-        self.setWindowFlags(Qt.FramelessWindowHint |
-                            Qt.WindowMinMaxButtonsHint)
-        self.windowEffect.addWindowAnimation(self.winId())
-
-        if "Windows-7" in platform():
-            self.windowEffect.addShadowEffect(self.winId())
-            self.windowEffect.setAeroEffect(self.winId())
-        else:
-            self.windowEffect.setAcrylicEffect(self.winId())
-            if sys.getwindowsversion().build >= 22000:
-                self.windowEffect.addShadowEffect(self.winId())
-
-        self.setStyleSheet("background:transparent")
-
-    def nativeEvent(self, eventType, message):
-        """ Handle the Windows message """
-        msg = MSG.from_address(message.__int__())
-
-        # handle Alt+F4
-        if msg.message == win32con.WM_SYSKEYDOWN and msg.wParam == win32con.VK_F4:
-            self.__closedByKey = True
-            QApplication.sendEvent(self, QCloseEvent())
-            return False, 0
-
-        return super().nativeEvent(eventType, message)
-
-    def closeEvent(self, e):
-        if not self.__closedByKey or QApplication.quitOnLastWindowClosed():
-            self.__closedByKey = False
-            return super().closeEvent(e)
-
-        # system tray icon
-        self.__closedByKey = False
-        self.hide()
