--- pyqtribbon/mainwindow.py
@@ -1,8 +1,0 @@
-from .framelesswindow import FramelessWindow
-
-
-class RibbonMainWindow(FramelessWindow):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.titleBar.raise_()
