+++ pyqtribbon/utils/linux_utils.py
